Start title cards
We can use these, or not, as needed. I broke up the text so it was not so overwhelming. I have an alternate screen for screen02. The files are in order by the first number: 01, 02, 03 ect. I am assuming we are not going to be using the instruction page if we are doing analog but I'll put it here anyway just encase.